Remove killing of old unicorn after restart.
It is somewhat conventional to have this in your unicorn config file in
before_fork block

  # Quit the old unicorn process
  old_pid = "#{server.config[:pid]}.oldbin"
  if File.exists?(old_pid) && server.pid != old_pid
    begin
      Process.kill("QUIT", File.read(old_pid).to_i)
    rescue Errno::ENOENT, Errno::ESRCH
      # someone else did our job for us
    end
  end

Moreover, there might be different variations with incremental phase out
of old master, etc.

So it's better to have full control of phase out strategy in your
unicorn config rather than putting this on capistrano task.